# mamikost/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Lorem Ipsum",
        "title": "Mamikost",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ivanmuhammad977@gmail.com",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
...
        "/order/detail/": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "AddOrderDetail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderDetails"
                ],
                "summary": "AddOrderDetail",
                "parameters": [
                    {
                        "description": "Order detail request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddOrderDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/order/detail/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "RemoveItemFromOrder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderDetails"
                ],
                "summary": "RemoveItemFromOrder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
...
        "/repro/": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "GetAllRentProperties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RentProperty"
                ],
                "summary": "GetAllRentProperties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "CreateRentProperty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RentProperty"
                ],
                "summary": "CreateRentProperty",
                "parameters": [
                    {
                        "description": "Create Rent Property Request",
                        "name": "property",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRentPropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
...
        "/repro/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "GetRentPropertyByID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RentProperty"
                ],
                "summary": "GetRentPropertyByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "UpdateRentProperty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RentProperty"
                ],
                "summary": "UpdateRentProperty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Rent Property Request",
                        "name": "property",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRentPropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "DeleteRentProperty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RentProperty"
                ],
                "summary": "DeleteRentProperty",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddOrderDetailRequest": {
            "type": "object",
            "properties": {
                "orpd_orpo_id": {
                    "type": "integer"
                },
                "orpd_price": {
                    "type": "number"
                },
                "orpd_qty_unit": {
                    "type": "integer"
                },
                "orpd_repo_id": {
                    "type": "integer"
                },
                "orpd_total_price": {
                    "type": "number"
                }
            }
        },
...
        "models.CreateRentPropertyRequest": {
            "type": "object",
            "required": [
                "repo_cate_id",
                "repo_name",
                "repo_price"
            ],
            "properties": {
                "repo_cate_id": {
                    "type": "integer"
                },
                "repo_desc": {
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "repo_price": {
                    "type": "number"
                }
            }
        },
...
        "models.UpdateRentPropertyRequest": {
            "type": "object",
            "required": [
                "repo_cate_id",
                "repo_name",
                "repo_price"
            ],
            "properties": {
                "repo_cate_id": {
                    "type": "integer"
                },
                "repo_desc": {
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "repo_price": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}